refactor: mergeThemes function
