+++ src/utils/helpers/checkMinMax.ts
