+++ src/utils/helpers/index.ts
@@ -1,4 +1,4 @@
-export { checkMaxDate, checkMinDate } from './checkDateRange';
+export { checkMaxDate, checkMinDate } from './checkMinMax';
 export { mergeThemes } from './mergeThemes';
 export { stringToDate } from './stringToDate';
 export { timestampToDateFormat } from './timestampToDateFormat';
