+++ src/utils/helpers/mergeThemes.ts
@@ -1,21 +1,20 @@
 import { ThemeType } from '@/types';
 
-type ObjType = { [key: string]: string | ObjType };
+type ThemeObjType = { [key: string]: string | ThemeObjType };
 
 export function mergeThemes(theme: ThemeType, customTheme?: ThemeType): ThemeType {
   if (!customTheme) return theme;
 
-  /* eslint-disable no-restricted-syntax */
-  const updateDefaultTheme = (obj: ObjType, secondObj: ObjType) => {
-    for (const key of Object.keys(obj)) {
-      if (typeof obj[key] === 'object' && typeof secondObj[key] === 'object') {
-        updateDefaultTheme(obj[key] as ObjType, secondObj[key] as ObjType);
+  const updateDefaultTheme = (themeObj: ThemeObjType, customThemeObj: ThemeObjType) => {
+    Object.keys(themeObj).forEach((key) => {
+      if (typeof themeObj[key] === 'object' && typeof customThemeObj[key] === 'object') {
+        updateDefaultTheme(themeObj[key] as ThemeObjType, customThemeObj[key] as ThemeObjType);
       } else {
-        obj[key] = secondObj?.[key] ?? obj[key];
+        themeObj[key] = customThemeObj?.[key] ?? themeObj[key];
       }
-    }
+    });
 
-    return obj;
+    return themeObj;
   };
 
   return updateDefaultTheme(theme, customTheme);
